Moved classes from MissionComponents.py into their own files
