+++ src/model/components/__init__.py
@@ -1,0 +1,18 @@
+"""
+components
+==========
+
+Provides all the custom objects that the data model needs
+"""
+
+from .Cargo import Cargo
+from .Clearance import Clearance
+from .Deadline import Deadline
+from .Destination import Destination
+from .Illegal import Illegal
+from .Log import Log
+from .Passengers import Passengers
+from .Source import Source
+from .Stopover import Stopover
+from .Trigger import Trigger
+from .TriggerCondition import TriggerCondition
